requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
